+++ app/ReactFlowComponents/UFComponents/components/TAccordion/TAccordion.tsx
@@ -2,7 +2,7 @@
 import React from "react";
 import { Accordion, AccordionItem } from "@nextui-org/react";
 
-const Taccordion = ({}) => {
+const Taccordion = (props) => {
   const defaultContent =
     "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.";
