+++ app/ReactFlowComponents/UFComponents/components/TAutocomplete/TAutocomplete.tsx
@@ -1,20 +1,17 @@
-"use client"
+"use client";
 import React from "react";
 import { Autocomplete, AutocompleteItem } from "@nextui-org/react";
 
 const animals = [
-  { value: 'cat', label: 'Cat' },
-  { value: 'dog', label: 'Dog' },
-  { value: 'bird', label: 'Bird' },
+  { value: "cat", label: "Cat" },
+  { value: "dog", label: "Dog" },
+  { value: "bird", label: "Bird" },
 ];
 
-const Tautocomplete = () => {
+const Tautocomplete = (props) => {
   return (
     <div className="flex w-full flex-wrap md:flex-nowrap gap-4">
-      <Autocomplete 
-        label="Select an animal" 
-        className="max-w-xs" 
-      >
+      <Autocomplete label="Select an animal" className="max-w-xs">
         {animals.map((animal) => (
           <AutocompleteItem key={animal.value} value={animal.value}>
             {animal.label}
@@ -35,6 +32,6 @@ const Tautocomplete = () => {
       </Autocomplete>
     </div>
   );
-}
+};
 
 export default Tautocomplete;
