+++ app/ReactFlowComponents/UFComponents/components/TBadge/TBadge.tsx
@@ -1,9 +1,8 @@
 import React from "react";
-import {Badge,Avatar} from "@nextui-org/react";
-const Tbadge= () => {
-    return (
-        <React.Fragment>
-        
+import { Badge, Avatar } from "@nextui-org/react";
+const Tbadge = (props) => {
+  return (
+    <React.Fragment>
       <Badge content="5" color="primary">
         <Avatar
           radius="md"
@@ -11,7 +10,7 @@ const Tbadge= () => {
           src="https://i.pravatar.cc/300?u=a042581f4e29026709d"
         />
       </Badge>
-      </React.Fragment>
-    );
-  }
-  export default Tbadge
+    </React.Fragment>
+  );
+};
+export default Tbadge;
