+++ app/ReactFlowComponents/UFComponents/components/TCheckbox/TCheckbox.tsx
@@ -1,9 +1,6 @@
 import React from "react";
-import {Checkbox} from "@nextui-org/react";
-const Tcheckbox = () => {
-    return (
-      <Checkbox defaultSelected>Option</Checkbox>
-    );
-  }
-  export default Tcheckbox
-   
+import { Checkbox } from "@nextui-org/react";
+const Tcheckbox = (props) => {
+  return <Checkbox defaultSelected>Option</Checkbox>;
+};
+export default Tcheckbox;
