+++ app/ReactFlowComponents/UFComponents/components/TCode/TCode.tsx
@@ -1,9 +1,7 @@
 import React from "react";
-import {Code} from "@nextui-org/react";
+import { Code } from "@nextui-org/react";
 
-const Tcode =()=> {
-  return (
-    <Code>npm install @nextui-org/react</Code>
-  );
-}
-export default Tcode
+const Tcode = (props) => {
+  return <Code>npm install @nextui-org/react</Code>;
+};
+export default Tcode;
