+++ app/ReactFlowComponents/UFComponents/components/TDivider/TDivider.tsx
@@ -1,11 +1,13 @@
 import React from "react";
-import {Divider} from "@nextui-org/react";
-const Tdivider =()=>{
-    return(
-        <div className="max-w-md">
+import { Divider } from "@nextui-org/react";
+const Tdivider = (props) => {
+  return (
+    <div className="max-w-md">
       <div className="space-y-1">
         <h4 className="text-medium font-medium">NextUI Components</h4>
-        <p className="text-small text-default-400">Beautiful, fast and modern React UI library.</p>
+        <p className="text-small text-default-400">
+          Beautiful, fast and modern React UI library.
+        </p>
       </div>
       <Divider className="my-4" />
       <div className="flex h-5 items-center space-x-4 text-small">
@@ -16,7 +18,6 @@ const Tdivider =()=>{
         <div>Source</div>
       </div>
     </div>
-
-    )
-}
-export default Tdivider
+  );
+};
+export default Tdivider;
