+++ app/ReactFlowComponents/UFComponents/components/Tcard/Tcard.tsx
@@ -1,15 +1,14 @@
 import React from "react";
-import {Card, CardHeader, CardBody, CardFooter} from "@nextui-org/react";
-const Tcard = () => {
-
-    return (
-        <React.Fragment>
+import { Card, CardHeader, CardBody, CardFooter } from "@nextui-org/react";
+const Tcard = (props) => {
+  return (
+    <React.Fragment>
       <Card>
         <CardBody>
           <p>Make beautiful websites regardless of your design experience.</p>
         </CardBody>
       </Card>
-      </React.Fragment>
-    );
-  }
-  export default Tcard 
+    </React.Fragment>
+  );
+};
+export default Tcard;
