+++ app/ReactFlowComponents/UFComponents/components/Tchip/TChip.tsx
@@ -1,9 +1,6 @@
 import React from "react";
-import {Chip} from "@nextui-org/react";
-const Tchip= ()=>{
-    return(
-        <Chip>Chip</Chip>
-
-    )
-}
-export default Tchip
+import { Chip } from "@nextui-org/react";
+const Tchip = (props) => {
+  return <Chip>Chip</Chip>;
+};
+export default Tchip;
